requirements.txt
Django==2.2.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.1.0
django-crispy-forms==1.8.0
django-datatables-view==1.19.0
django-fontawesome==1.0
django-tables2==2.1.1
django-tables2-bootstrap4==2.0.0b5
django-user-language-middleware==0.0.3
djangorestframework==3.10.3
djangorestframework-datatables-editor==0.3.2
pytz==2019.3
PyYAML==5.1.2
sqlparse==0.3.0
